added a working version using google sheet api
this is an (ugly) implementation that uses the google sheet api and javascript. 
it is badly commented and has no css, but it functions and connects to a sheet, adding the values where needed.
I think that we should use this or something similar as the core of the final project.